Create LICENCE.md
Permission is hereby granted, free of charge, to any person obtaining a copy of the work and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions: The software is provided "as is", without warranty of any kind.
In no event shall the authors or copyright holders be liable for any claim, damages or other liability, whether in an action of contract, tort or otherwise, arising from, out of or in connection with the software or the use or other dealings in the software.
The visual publications: “Money, credit and central bank interest rates” and “Inflation” are two interactive publications built upon and made available by Eurostat, see the Eurostat Disclaimer https://ec.europa.eu/info/legal-notice_en. The ECB/Eurostat does not take responsibility for any replication of the content of the Software, or any other form of redistribution.